requirements.txt
langchain
langgraph
langgraph-cli[inmem]
pydantic
dspy-ai
huggingface_hub
faiss-gpu-cu12
transformers
torch
sentence-transformers
numpy
pandas
scikit-learn
networkx
matplotlib
pyvis
